+++ majortom_scripting/__init__.py
@@ -1,2 +1,3 @@
 from majortom_scripting.scripting_api import ScriptingAPI
+from majortom_scripting.modeling_api import ModelingAPI
 name = "majortom_scripting"