+++ majortom_scripting/models/groundstation.py
@@ -1,0 +1,7 @@
+from majortom_scripting.models.system import System
+
+class GroundStation(System):
+
+  def update_location(self, lat, long):
+    #tbd
+    pass