+++ majortom_scripting/models/satellite.py
@@ -1,0 +1,6 @@
+from majortom_scripting.models.system import System
+
+class Satellite(System):
+
+  def update_tle(self, tle1, tle2):
+    pass